requirements.txt
langchain
langchain-community
dashscope
chromadb
faiss-cpu>=1.8.0
numpy<2.0.0
tiktoken
python-dotenv
streamlit
markdown
python-docx
PyPDF2
pdfplumber
unstructured[local-inference]
python-magic-bin
# gRPC dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0